fix: delete cascade when deleting user
When deleting a user, a user's feedbacks in the database would not get
deleted.  It instead would have its username attribute set to null.
Having an "on delete cascade" setting on the username foreign key in
the Feedback class is not enough.  That only sets the constraint on the
database side.

A cascade option with "all, delete-orphan" for the relationship
"feedbacks" in the User class is needed to delete the feedbacks
belonging to a user when a user is deleted.  This piece of code will
apply the cascading delete on the server/SQLAlchemy side.